// src/app/shared/model/recipient.model.ts
export interface Recipient {
    user_id: string;
    email: string;
    user_name: string;
    receptionist_email: string;
    is_director_or_board: string;
    user_surname: string;
}

export interface RecipientDTO{ 
    user_id: string;
    email: string;
    user_name: string;
    receptionist_email: string;
    is_director_or_board: string;
    user_surname: string;
    
} 