// src/app/shared/model/profile.model.ts
export interface Profile {
    person_id?: string;
    person_number?: string;
    position?: string;
    position_desc?: string;
    position_desc_lang?: string;
    job_id?: string;
    job_description?: string;
    job_description_lang?: string;
    job_risk_code?: string;
    job_risk_description?: string;
    userid_1?: string;
    userid_sap?: string;
    userid_2?: string;
    userid_3?: string;
    userid_4?: string;
    u_org?: string;
    u_org_desc?: string;
    cost_center?: string;
    cost_center_desc?: string;
    last_name?: string;
    second_name?: string;
    first_name?: string;
    birthday?: string;
    taxid?: string;
    residence_street?: string;
    residence_street_add?: string;
    residence_zip?: string;
    residence_txmun?: string;
    residence_state?: string;
    residence_state_desc?: string;
    residence_country_desc?: string;
    last_name_l?: string;
    first_name_l?: string;
    role?: string;
    role_desc?: string;
    gender?: string;
    language?: string;
    nationality?: string;
    nationality_desc?: string;
    birth_country?: string;
    birth_place_id?: string;
    birth_place_desc?: string;
    starting_date?: string;
    termination_date?: string;
    company?: string;
    company_desc?: string;
    pers_area?: string;
    pers_area_desc?: string;
    subarea?: string;
    subarea_desc?: string;
    country?: string;
    country_desc?: string;
    my_hrbp?: string;
    my_hrbp_local?: string;
    perimeter?: string;
    perimeter_desc?: string;
    perimeter_short?: string;
    state?: string;
    emp_group?: string;
    emp_group_desc?: string;
    emp_subgroup?: string;
    emp_subgroup_desc?: string;
    contract_type?: string;
    contract_type_des?: string;
    working_hs?: string;
    level?: string;
    emp_level?: string;
    emp_category?: string;
    wc_code?: string;
    wc_address?: string;
    wc_cap?: string;
    wc_city?: string;
    wc_region?: string;
    wc_country?: string;
    person_id_man_1?: string;
    person_num_man_1?: string;
    type_man_1?: string;
    relat_man_1?: string;
    relat_desc_man_1?: string;
    person_id_man_2?: string;
    person_num_man_2?: string;
    type_man_2?: string;
    relat_man_2?: string;
    relat_desc_man_2?: string;
    person_id_man_3?: string;
    person_num_man_3?: string;
    type_man_3?: string;
    relat_man_3?: string;
    relat_desc_man_3?: string;
    person_id_man_4?: string;
    person_num_man_4?: string;
    type_man_4?: string;
    relat_man_4?: string;
    relat_desc_man_4?: string;
    person_id_man_5?: string;
    person_num_man_5?: string;
    type_man_5?: string;
    relat_man_5?: string;
    relat_desc_man_5?: string;
    person_id_man_6?: string;
    person_num_man_6?: string;
    type_man_6?: string;
    relat_man_6?: string;
    relat_desc_man_6?: string;
    person_id_adm_man?: string;
    person_num_adm_man?: string;
    person_id_top_man?: string;
    person_num_top_man?: string;
    person_num_adm_loc_man?: string;
    person_id_adm_loc_man?: string;
    flag_head?: string;
    person_id_bp?: string;
    email?: string;
    tel_1?: string;
    tel_2?: string;
    tel_3?: string;
    mobile_1?: string;
    mobile_2?: string;
    tel_short_1?: string;
    tel_short_2?: string;
    tel_short_3?: string;
    tel_secr_1?: string;
    tel_secr_2?: string;
    tel_secr_3?: string;
    net_salary?: string;
    net_salary_curr?: string;
    last_annual_bonus?: string;
    last_annual_bonus_curr?: string;
    current_ral?: string;
    current_ral_curr?: string;
    per_final_result?: string;
    per_final_result_des?: string;
    pred_attr?: string;
    flpa_key?: string;
    flpa_long_text?: string;
    attr_prev_cod?: string;
    attr_prev_desc?: string;
    attr_prev_abbr?: string;
    bus_line_cod_1?: string;
    bus_line_desc_1?: string;
    bus_line_cod_2?: string;
    bus_line_desc_2?: string;
    bus_line_cod_3?: string;
    bus_line_desc_3?: string;
    bus_line_cod_4?: string;
    bus_line_desc_4?: string;
    bus_line_cod_5?: string;
    bus_line_desc_5?: string;
    serv_cod_1?: string;
    serv_desc_1?: string;
    serv_cod_2?: string;
    serv_desc_2?: string;
    serv_cod_3?: string;
    serv_desc_3?: string;
    staff_cod_1?: string;
    staff_desc_1?: string;
    staff_cod_2?: string;
    staff_desc_2?: string;
    org_chart_di?: string;
    central_person?: string;
    pernr_exp?: string;
    company_exp?: string;
    company_desc_exp?: string;
    country_exp?: string;
    country_desc_exp?: string;
    userid_1_exp?: string;
    userid_1_exp_sap?: string;
    email_exp?: string;
    mobile_1_exp?: string;
    mobile_2_exp?: string;
    net_salary_exp?: string;
    net_salary_curr_exp?: string;
    last_annual_bonus_exp?: string;
    last_annual_bonus_curr_exp?: string;
    current_ral_exp?: string;
    current_ral_curr_exp?: string;
    start_date_imp?: string;
    end_date_imp?: string;
    onboarding?: string;
}